+++ lib/main.dart
@@ -1,97 +1,6 @@
 import 'package:flutter/material.dart';
-
-import 'model/database_handler.dart';
-import 'model/user.dart';
+import 'view/my_app.dart';
 
 void main() {
   runApp(const MyApp());
-}
-
-/// adding first branch
-class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter smimple database',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: const MyHomePage(title: 'Flutter smimple database'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({Key? key, required this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  late DatabaseHandler handler;
-
-  @override
-  void initState() {
-    Future<int> addUsers() async {
-      User firstUser = User(name: "Mana", location: "Nabagram");
-      User secondUser = User(name: "Babu", location: "Nabagram");
-      User thirdUser = User(name: "Pata", location: "Nabagram");
-      List<User> listOfUsers = [
-        firstUser,
-        secondUser,
-        thirdUser,
-      ];
-      return await handler.insertUser(listOfUsers);
-    }
-
-    super.initState();
-    handler = DatabaseHandler();
-    handler.initializeDB().whenComplete(() async {
-      await addUsers();
-      setState(() {});
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: FutureBuilder(
-        future: handler.retrieveUsers(),
-        builder: (BuildContext context, AsyncSnapshot<List<User>> snapshot) {
-          if (snapshot.hasData) {
-            return ListView.builder(
-              itemCount: snapshot.data?.length,
-              itemBuilder: (BuildContext context, int index) {
-                return Card(
-                  child: ListTile(
-                    key: ValueKey<int>(snapshot.data![index].id!),
-                    contentPadding: const EdgeInsets.all(8.0),
-                    title: Text(
-                      snapshot.data![index].name,
-                      style: Theme.of(context).textTheme.headline3,
-                    ),
-                    subtitle: Text(
-                      snapshot.data![index].location.toString(),
-                      style: Theme.of(context).textTheme.headline5,
-                    ),
-                  ),
-                );
-              },
-            );
-          } else {
-            return const Center(child: CircularProgressIndicator());
-          }
-        },
-      ),
-    );
-  }
 }
